% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaugePlot.R
\name{gaugePlot}
\alias{gaugePlot}
\title{Get a gauge plot for use in the app.}
\usage{
gaugePlot(
  translator,
  rating = "Empty",
  type = "practices",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
Get a gauge plot for use in the app.
}
